move everything to btrfs2s3._internal
the intent is to make sure no one depends on the programmatic interface,
as I don't intend to make it public yet
