+++ src/btrfs2s3/_internal/assessor.py
@@ -12,14 +12,14 @@ import arrow
 import btrfsutil
 from btrfsutil import SubvolumeInfo
 
+from btrfs2s3._internal.resolver import Flags
+from btrfs2s3._internal.resolver import KeepMeta
+from btrfs2s3._internal.resolver import resolve
+from btrfs2s3._internal.s3 import iter_backups
+from btrfs2s3._internal.thunk import Thunk
+from btrfs2s3._internal.thunk import ThunkArg
 from btrfs2s3._internal.util import mksubvol
 from btrfs2s3._internal.util import SubvolumeFlags
-from btrfs2s3.resolver import Flags
-from btrfs2s3.resolver import KeepMeta
-from btrfs2s3.resolver import resolve
-from btrfs2s3.s3 import iter_backups
-from btrfs2s3.thunk import Thunk
-from btrfs2s3.thunk import ThunkArg
 
 if TYPE_CHECKING:
     from pathlib import Path
@@ -27,10 +27,10 @@ if TYPE_CHECKING:
 
     from mypy_boto3_s3.client import S3Client
 
-    from btrfs2s3 import resolver
-    from btrfs2s3.action import Actions
-    from btrfs2s3.backups import BackupInfo
-    from btrfs2s3.preservation import Policy
+    from btrfs2s3._internal import resolver
+    from btrfs2s3._internal.action import Actions
+    from btrfs2s3._internal.backups import BackupInfo
+    from btrfs2s3._internal.preservation import Policy
 
 
 @dataclass
