+++ src/btrfs2s3/_internal/stream_uploader.py
