+++ tests/action/create_backup_test.py
@@ -6,7 +6,7 @@ import subprocess
 from typing import TYPE_CHECKING
 
 from botocore.exceptions import ClientError
-from btrfs2s3.action import create_backup
+from btrfs2s3._internal.action import create_backup
 import btrfsutil
 import pytest
