+++ tests/action/rename_snapshot_test.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from btrfs2s3.action import rename_snapshot
+from btrfs2s3._internal.action import rename_snapshot
 import btrfsutil
