+++ tests/commands/update/describe_time_span_test.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
 import arrow
-from btrfs2s3.commands.update import describe_time_span
-from btrfs2s3.zoneinfo import get_zoneinfo
+from btrfs2s3._internal.commands.update import describe_time_span
+from btrfs2s3._internal.zoneinfo import get_zoneinfo
 import pytest
 from rich.text import Span
