+++ tests/commands/update/print_actions_test.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 from pathlib import Path
 from typing import TYPE_CHECKING
 
-from btrfs2s3.action import Actions
-from btrfs2s3.commands.update import print_actions
+from btrfs2s3._internal.action import Actions
+from btrfs2s3._internal.commands.update import print_actions
 import pytest
 
 if TYPE_CHECKING:
