+++ tests/resolver/keep_snapshot_and_backup_for_time_span_test.py
@@ -5,17 +5,17 @@ from typing import TYPE_CHECKING
 from uuid import uuid4
 
 import arrow
+from btrfs2s3._internal.preservation import Params
+from btrfs2s3._internal.preservation import Policy
+from btrfs2s3._internal.resolver import _Resolver
+from btrfs2s3._internal.resolver import Flags
+from btrfs2s3._internal.resolver import KeepBackup
+from btrfs2s3._internal.resolver import KeepMeta
+from btrfs2s3._internal.resolver import KeepSnapshot
+from btrfs2s3._internal.resolver import Reasons
+from btrfs2s3._internal.resolver import Result
 from btrfs2s3._internal.util import backup_of_snapshot
 from btrfs2s3._internal.util import mksubvol
-from btrfs2s3.preservation import Params
-from btrfs2s3.preservation import Policy
-from btrfs2s3.resolver import _Resolver
-from btrfs2s3.resolver import Flags
-from btrfs2s3.resolver import KeepBackup
-from btrfs2s3.resolver import KeepMeta
-from btrfs2s3.resolver import KeepSnapshot
-from btrfs2s3.resolver import Reasons
-from btrfs2s3.resolver import Result
 import pytest
 
 if TYPE_CHECKING:
